added bitmap converter class
