+++ app/src/main/java/com/werelit/neurolls/neurolls/SearchMediaActivity.java
@@ -32,6 +32,7 @@ import com.android.volley.VolleyError;
 import com.igdb.api_android_java.callback.onSuccessCallback;
 import com.igdb.api_android_java.model.APIWrapper;
 import com.igdb.api_android_java.model.Parameters;
+import com.werelit.neurolls.neurolls.network.BitmapConverter;
 import com.werelit.neurolls.neurolls.network.ConnectGameDB;
 import com.werelit.neurolls.neurolls.model.Book;
 import com.werelit.neurolls.neurolls.model.Film;
@@ -436,7 +437,7 @@ public class SearchMediaActivity extends AppCompatActivity implements LoaderMana
         bundle.putString(MediaKeys.MEDIA_GENRE_KEY, media.getmMediaGenre());
         bundle.putString(MediaKeys.MEDIA_YEAR_KEY, media.getmMediaYear());
         if(media.getThumbnailBmp() != null)
-            bundle.putString(MediaKeys.MEDIA_IMAGE_KEY, bitmapToString(scaleDownBitmap(media.getThumbnailBmp(), SCALE, this)));
+            bundle.putString(MediaKeys.MEDIA_IMAGE_KEY, BitmapConverter.bitmapToString(BitmapConverter.scaleDownBitmap(media.getThumbnailBmp(), SCALE, this)));
         bundle.putInt(MediaKeys.MEDIA_CATEGORY_KEY, getCategoryCode(media));
     }
 
@@ -594,7 +595,7 @@ public class SearchMediaActivity extends AppCompatActivity implements LoaderMana
                 int n = Integer.parseInt(currentArchived);
                 film.setArchived((n == 1)? true : false);
                 film.setNotifSettings(currentNotif);
-                film.setThumbnailBmp(stringToBitMap(currentImage));
+                film.setThumbnailBmp(BitmapConverter.stringToBitMap(currentImage));
                 //Log.wtf(LOG_TAG, "CURRENT ARCHIVED: " + n);
                 mediaList.add(0, film);
             }
@@ -805,38 +806,38 @@ public class SearchMediaActivity extends AppCompatActivity implements LoaderMana
         }
     }
 
-    public String bitmapToString(Bitmap bitmap){
-        ByteArrayOutputStream baos = new  ByteArrayOutputStream();
-        bitmap.compress(Bitmap.CompressFormat.PNG,100, baos);
-        byte [] b = baos.toByteArray();
-        String strBitmap = Base64.encodeToString(b, Base64.DEFAULT);
-        return strBitmap;
-    }
-
-    /**
-     * @param encodedString
-     * @return bitmap (from given string)
-     */
-    public Bitmap stringToBitMap(String encodedString){
-        try {
-            byte [] encodeByte = Base64.decode(encodedString, Base64.DEFAULT);
-            Bitmap bitmap = BitmapFactory.decodeByteArray(encodeByte, 0, encodeByte.length);
-            return bitmap;
-        } catch(Exception e) {
-            e.getMessage();
-            return null;
-        }
-    }
-
-    public static Bitmap scaleDownBitmap(Bitmap photo, int newHeight, Context context) {
-
-        final float densityMultiplier = context.getResources().getDisplayMetrics().density;
-
-        int h= (int) (newHeight*densityMultiplier);
-        int w= (int) (h * photo.getWidth()/((double) photo.getHeight()));
-
-        photo=Bitmap.createScaledBitmap(photo, w, h, true);
-
-        return photo;
-    }
+//    public String bitmapToString(Bitmap bitmap){
+//        ByteArrayOutputStream baos = new  ByteArrayOutputStream();
+//        bitmap.compress(Bitmap.CompressFormat.PNG,100, baos);
+//        byte [] b = baos.toByteArray();
+//        String strBitmap = Base64.encodeToString(b, Base64.DEFAULT);
+//        return strBitmap;
+//    }
+//
+//    /**
+//     * @param encodedString
+//     * @return bitmap (from given string)
+//     */
+//    public Bitmap stringToBitMap(String encodedString){
+//        try {
+//            byte [] encodeByte = Base64.decode(encodedString, Base64.DEFAULT);
+//            Bitmap bitmap = BitmapFactory.decodeByteArray(encodeByte, 0, encodeByte.length);
+//            return bitmap;
+//        } catch(Exception e) {
+//            e.getMessage();
+//            return null;
+//        }
+//    }
+//
+//    public static Bitmap scaleDownBitmap(Bitmap photo, int newHeight, Context context) {
+//
+//        final float densityMultiplier = context.getResources().getDisplayMetrics().density;
+//
+//        int h= (int) (newHeight*densityMultiplier);
+//        int w= (int) (h * photo.getWidth()/((double) photo.getHeight()));
+//
+//        photo=Bitmap.createScaledBitmap(photo, w, h, true);
+//
+//        return photo;
+//    }
 }
